Former-commit-id: 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee [formerly f65e8a7253c65b5d59e78ed57e3d73bf01b2e06f]]
Former-commit-id: be88c9494ecedbdb140a4995c6feec6b29c4cfee
Former-commit-id: 29677326c7a0b9dcf115b06d8cb2e0645a2c7614

+++ edexOsgi/com.raytheon.uf.edex.plugin.bufrmthdw/src/com/raytheon/uf/edex/plugin/bufrmthdw/decoder/MTHDWDataAdapter.java
@@ -52,6 +52,7 @@ import com.raytheon.uf.edex.pointdata.PointDataPluginDao;
  *                                     types.
  * Aug 30, 2013 2298       rjpeter     Make getPluginName abstract
  * May 14, 2014 2536       bclement    moved WMO Header to common, removed TimeTools usage
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -225,8 +226,8 @@ public class MTHDWDataAdapter extends BUFRPointDataAdapter<BufrMTHDWObs> {
                     IDecoderConstants.VAL_MISSING);
 
             SurfaceObsLocation location = new SurfaceObsLocation();
-            location.setLatitude(lat);
-            location.setLongitude(lon);
+            location.setLatitude((float) lat);
+            location.setLongitude((float) lon);
             location.generateCoordinateStationId();
             pointData.setLocation(location);
             // ****************************************
